Refactoring to build BFKit on Linux
